backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
segmentation-models-pytorch==0.3.3
scikit-image==0.22.0
python-dotenv==1.0.0
pydantic==2.5.0
